# docs/openapi.yaml
{
  "openapi": "3.0.0",
  "paths": {
...
    "/terminals": {
      "get": {
        "operationId": "TerminalsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Список терминалов"
          }
        },
        "security": [
          {
            "JWT-auth": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Получить список терминалов",
        "tags": [
          "Терминалы"
        ]
      },
      "post": {
        "operationId": "TerminalsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTerminalDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Терминал создан"
          }
        },
        "security": [
          {
            "JWT-auth": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Создать новый терминал",
        "tags": [
          "Терминалы"
        ]
      }
    },
    "/terminals/{id}": {
      "get": {
        "operationId": "TerminalsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Данные терминала"
          },
          "404": {
            "description": "Терминал не найден"
          }
        },
        "security": [
          {
            "JWT-auth": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Получить карточку терминала",
        "tags": [
          "Терминалы"
        ]
      },
      "patch": {
        "operationId": "TerminalsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTerminalDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Терминал обновлен"
          }
        },
        "security": [
          {
            "JWT-auth": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Обновить терминал",
        "tags": [
          "Терминалы"
        ]
      },
      "delete": {
        "operationId": "TerminalsController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Терминал удален"
          }
        },
        "security": [
          {
            "JWT-auth": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Удалить терминал",
        "tags": [
          "Терминалы"
        ]
      }
    },
    "/terminals/{id}/status": {
      "patch": {
        "description": "Ручное изменение статуса терминала (ACTIVE/INACTIVE). Используется для тестирования или административных целей. Требует роль ROOT или MANAGER.",
        "operationId": "TerminalsController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTerminalStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Статус терминала успешно обновлен",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "0bd492c9-5d78-428d-8ef3-3d5ece1296c6",
                    "macAddress": "FF:EE:DD:CC:BB:AA",
                    "status": "INACTIVE",
                    "shopId": "a96e7bad-d4d8-4559-b9b7-0e180fedd3e3",
                    "createdAt": "2025-10-19T22:45:05.685Z",
                    "updatedAt": "2025-10-19T22:46:56.829Z",
                    "shop": {
                      "id": "a96e7bad-d4d8-4559-b9b7-0e180fedd3e3",
                      "name": "Test Shop",
                      "address": "123 Test St"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Терминал не найден",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Terminal not found",
                    "error": "Not Found",
                    "statusCode": 404
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Обновить статус терминала вручную",
        "tags": [
          "Терминалы"
        ]
      }
    },
...
  },
  "info": {
    "title": "Shop Terminals Management API",
    "description": "",
    "version": "1.0",
    "contact": {

    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Введите JWT токен полученный при авторизации",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateTerminalDto": {
        "type": "object",
        "properties": {
          "macAddress": {
            "type": "string",
            "example": "AA:BB:CC:DD:EE:FF",
            "description": "MAC адрес терминала"
          },
          "shopId": {
            "type": "string",
            "example": "uuid-shop-id",
            "description": "ID магазина"
          }
        },
        "required": [
          "macAddress"
        ]
      },
      "UpdateTerminalDto": {
        "type": "object",
        "properties": {
          "macAddress": {
            "type": "string",
            "example": "AA:BB:CC:DD:EE:FF"
          },
          "status": {
            "type": "string",
            "example": "ACTIVE",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ]
          },
          "shopId": {
            "type": "string",
            "example": "uuid-shop-id"
          }
        }
      },
      "UpdateTerminalStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "ACTIVE",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ],
            "description": "Новый статус терминала"
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}